Update jsconfig.json so that autocomplete works for everyone.
The autocomplete features in various IDEs for JavaScript are based off
of jsconfig.json. In include, it has paths for TypeScript files that
describe types of functions that will be included, which it seems like
the p5 vscode extension automatically generates for
you. Unfortunately, it points to the path:

C:\Users\ephra\.vscode\extensions\samplavigne.p5-vscode-1.2.16\p5types\global.d.ts

which only works on Ephram's computer. I added the same folder that
the p5 extension uses to the actual repository, and added a path to it
for each jsconfig file. This way, everyone should be able to
autocomplete things. I kept the old line just so if I broke something
I didn't make anything actually worse than it was before.