Fix the actual speed camera
As it worked in the demo, if you had multiple speed cameras, and it
was within range of the first speed camera that you placed, but not
the second, the car wouldn't be slowed down. This is because when the
second camera's update() function ran, it would override whatever the
first one said. Now, there is a variable within the car object that
says whether it is currently being affected by a speed camera. The
variable is reset at the beginning of each frame, and set at the end
so that all the cars will be affected correctly.

I also added the range of the towers to the debug view, which should
be helpful when debugging and tweaking the code.

NOTE: This code will break if you change the cars to be updated before
the towers, which didn't exist as an issue before.